pyproject.toml
[tool.poetry]
name = "portfolio"
version = "0.1.0"
description = "Blog/portfolio"
authors = ["Oscar Peace <oscar@oscarcp.net>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^2.3.2"
gunicorn = "^21.2.0"
mistune = "^3.0.1"
flask-htmx = "^0.3.1"
markdown = "^3.4.4"
pillow = "^10.0.0"
flask-wtf = "^1.2.1"
flask-limiter = "^3.7.0"
argon2-cffi = "^23.1.0"
requests = "^2.32.3"
flask-compress = "^1.15"
sqlalchemy = "^2.0.36"
flask-sqlalchemy = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
